refactor: change files structure in components folder
