+++ src/pages/main/components/Editor/Editor.tsx
@@ -1,6 +1,6 @@
-import { QueryEditorWrapper } from './queryEditorWrapper/QueryEditorWrapper';
-import { QueryToolbar } from './QueryToolbar/QueryToolbar';
-import { Viewer } from './Viewer/Viewer';
+import { QueryEditorWrapper } from '../QueryEditorWrapper/QueryEditorWrapper';
+import { QueryToolbar } from '../QueryToolbar/QueryToolbar';
+import { Viewer } from '../Viewer/Viewer';
 import styles from './Editor.module.scss';
 
 export function Editor() {
