+++ src/pages/main/components/QueryEditor/QueryEditor.module.scss
