+++ src/pages/main/components/QueryEditor/QueryEditor.tsx
