+++ src/pages/main/components/QueryEditorWrapper/QueryEditorWrapper.module.scss
