+++ src/pages/main/components/QueryToolbar/QueryToolbar.module.scss
