+++ tsconfig.json
@@ -23,6 +23,7 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
     "noImplicitAny": true,
+    "forceConsistentCasingInFileNames": false,
   },
   "include": [
     ".eslintrc.cjs",
